kokulanK-Y2S2_Artificial_Intelligence_and_Machine_Learning_Project-main/model/requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
python-multipart
numpy==1.24.4
opencv-python-headless==4.8.1.78